v7 : Code refactor & Added iOS support

+++ src/main.py
@@ -2,24 +2,25 @@ import requests,json,os,shutil,sys,zipfile
 from appData import ApkCombo, Aptiode,TwtWeb
 from tqdm import tqdm
 from pprint import pprint
-from comman import DUMMY_FOLDER,MAIN_FOLDER,ZIP_FILE,EXTRACT_FOLDER,PKG_NAME,APP_NAME,new_file_name,old_file_name,DEBUG,manifest_file_name
+from comman import DUMMY_FOLDER,MAIN_FOLDER,ZIP_FILE,EXTRACT_FOLDER,PKG_NAME,APP_NAME,new_file_name,old_file_name,DEBUG,manifest_file_name,Platform,Releases,new_file_ipad_name,old_file_ipad_name
 from comman import writeJson
 
-VER = "v6.1 : Code refactor & repo name change : removed env loader : added exp flags and flag limit"
+VER = "v7 : Code refactor & Added iOS support"
 
 
 vername = "web"
-source = "manual"
-vercode = "estdyfutyf35"
+source = "web"
+vercode = ""
 down_link = ""
 
 vername = vername.lower()
 source = source.lower()
 vercode = vercode.lower()
 
-if os.path.exists(DUMMY_FOLDER):
-    shutil.rmtree(DUMMY_FOLDER)
-os.makedirs(MAIN_FOLDER)
+if not DEBUG:
+    if os.path.exists(DUMMY_FOLDER):
+        shutil.rmtree(DUMMY_FOLDER)
+    os.makedirs(MAIN_FOLDER)
 
 
 
@@ -38,27 +39,51 @@ def downloader(url,fileName=""):
     progress_bar.close()
 
 
-def unzipper():
+def unzipper(platform):
     try:
+        def extract(src,new_name):
+            try:
+                zip_obj.extract(src,path=EXTRACT_FOLDER)
+                os.rename(EXTRACT_FOLDER +src, new_name)
+                return True
+            except Exception as e:
+                print(str(e))
+                return False
+
         feature_file = False
         zip_obj = zipfile.ZipFile(ZIP_FILE, 'r')
         file_list = zip_obj.namelist()
-        apk_name = f'{PKG_NAME}.apk'
-        if apk_name in file_list:
-            zip_obj.extract(apk_name, path=EXTRACT_FOLDER)
-            zip_obj = zipfile.ZipFile(
-                EXTRACT_FOLDER+apk_name, 'r')
-            feature_file = True
+        
+        if platform == Platform.ANDROID.value:
+            FLAG_FOLDER ="res/raw"
+            FLAG_FILE = f"{FLAG_FOLDER}/feature_switch_manifest" 
+            apk_name = f'{PKG_NAME}.apk'
 
-        elif "res/raw/feature_switch_manifest" in file_list:
-            feature_file = True
+            if apk_name in file_list:
+                zip_obj.extract(apk_name, path=EXTRACT_FOLDER)
+                zip_obj = zipfile.ZipFile(
+                    EXTRACT_FOLDER+apk_name, 'r')
+                feature_file = True
 
-        if feature_file:
-            zip_obj.extract("res/raw/feature_switch_manifest",
-                            path=EXTRACT_FOLDER)
-            os.rename(EXTRACT_FOLDER +
-                      'res/raw/feature_switch_manifest', new_file_name)
-            return True
+            elif FLAG_FILE in file_list:
+                feature_file = True
+
+            if feature_file:
+                return extract(FLAG_FILE,new_file_name)
+        
+        elif platform == Platform.IOS.value:
+            rd1 = False;rd2 = False
+            FLAG_FOLDER = "Payload/Twitter.app"
+            FLAG_FILE = f"{FLAG_FOLDER}/fs_embedded_defaults_ipad_production.json" 
+            if FLAG_FILE in file_list:
+                rd1 = extract(FLAG_FILE,new_file_ipad_name)
+            
+            FLAG_FILE = f"{FLAG_FOLDER}/fs_embedded_defaults_production.json" 
+            if FLAG_FILE in file_list:
+                rd2 = extract(FLAG_FILE,new_file_name)
+            
+            return rd1 and rd2
+        return False
     except Exception as e:
         print(str(e))
 
@@ -119,19 +144,19 @@ def downTwt2(typ):
 
 
 def main():
+    # global vername,source,vercode,down_link
     if len(sys.argv)>1:
         vername = sys.argv[1]
         source = sys.argv[2]
         vercode = sys.argv[3]
         down_link = sys.argv[4]
-    try:
-        
+    try:    
         hash_value = False
-        typ = "stable" if "release" in vername else "beta" if "beta" in vername else "alpha" if "alpha" in vername else "web"
+        typ = Releases.WEB.value if "web" in vername else Releases.BETA.value if "beta" in vername else Releases.ALPHA.value if "alpha" in vername else Releases.STABLE.value
+        platform = Platform.WEB.value if "web" in source else Platform.IOS.value if "ios" in source else Platform.ANDROID.value
         down_data = [False,False,False] #vername,vercode,downLink
-      
-  
-        if typ=="web":
+        
+        if platform==Platform.WEB.value:
             twt = TwtWeb()
             sha,version = twt.version()
             hash_value = sha
@@ -144,9 +169,27 @@ def main():
             shutil.copy(existsing_flag_file, new_file_name)
             down_data = [typ,version,False]
         
-        
+        elif platform==Platform.IOS.value:
+            fileName = "x.ipa"
+            vercode = ""
+            downloader(down_link,fileName)
+            s = unzipper(platform)
+            if not s:
+                return False
             
-        else:
+            existsing_flag_file = f'flags_iphone_{typ}.json'
+            os.rename(existsing_flag_file, old_file_name)
+            # os.remove(existsing_flag_file)
+            shutil.copy(new_file_name, existsing_flag_file)
+
+            existsing_flag_file = f'flags_ipad_{typ}.json'
+            os.rename(existsing_flag_file, old_file_ipad_name)
+            # os.remove(existsing_flag_file)
+            shutil.copy(new_file_ipad_name, existsing_flag_file)
+
+            down_data = [vername,vercode,down_link]
+            
+        elif platform==Platform.ANDROID.value:
             if source=="manual":
                 fileName = f"{vername}.apk"
                 downloader(down_link,fileName)
@@ -159,16 +202,16 @@ def main():
                 if not down_data[0]: 
                     return False
             
-            existsing_flag_file = f'flags_{typ}.json'
+            existsing_flag_file = f'flags_android_{typ}.json'
             shutil.copyfile(existsing_flag_file, old_file_name)
 
-            s = unzipper()
+            s = unzipper(platform)
             if not s:
                 return False
             os.remove(existsing_flag_file)
             shutil.copy(new_file_name, existsing_flag_file)
             
-        d = {'version_name': down_data[0],'vercode': down_data[1],'hash':hash_value,'download_link':down_data[2]}
+        d = {'version_name': down_data[0],'vercode': down_data[1],'hash':hash_value,'download_link':down_data[2],'os':platform}
         writeJson(manifest_file_name,d)
 
         return True
